scripts/requirements.txt
google-generativeai
supabase
requests
beautifulsoup4
tqdm